Intro to Declarative Macros
Dec Macros are a way to do code replacement, by use of syntax matching.
Here during the declaration, the matching syntaxes are placed which are
called based on the caller's passed arguements.

As Dec Macros are direct code substitions, They can't be expanded to
a bunch of statements, instead they have to return a block of expressions
which when executres return a single statement in the Vec case, as its
directly assigned at the caller.

Proc Macros are a more advanced use-case, and allow you to change even the
input syntax, and `macro_rules!` uses that internally.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>