Use Pattern Marching to get unlimited arugements
As we can't write a macro pattern match for each additional arguement,
we use pattern matching regex-like thing to make it expandable,
This is done by using wrapping the input arg into a $(x),+ which means
allow multiple of x's, where x is $element:expr

As we also need to do the same internally, that is perform a push for
each arguement, we do the same i.e $(vs.push($element);)*
We also tackle the trailing comma problem.

We also allow Vec![;] here, and learn a imp lesson of evaulating the
arguement once and cloning it to use it multiple times, as the passed
arg can be a single evaluation item i.e a x.take().unwrap(), etc.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>